+++ src/test/java/org/folio/am/support/KafkaEventAssertions.java
@@ -3,7 +3,7 @@ package org.folio.am.support;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.folio.am.integration.kafka.DiscoveryPublisher.DISCOVERY_DESTINATION;
 import static org.folio.common.utils.CollectionUtils.mapItems;
-import static org.folio.integration.kafka.KafkaUtils.getEnvTopicName;
+import static org.folio.integration.kafka.producer.KafkaUtils.getEnvTopicName;
 import static org.folio.test.FakeKafkaConsumer.getEvents;
 import static org.testcontainers.shaded.org.awaitility.Durations.ONE_MINUTE;
 import static org.testcontainers.shaded.org.awaitility.Durations.ONE_SECOND;